+++ guillotina/events.py
@@ -1,10 +1,13 @@
 from guillotina.component.interfaces import IObjectEvent
 from guillotina.interfaces import IBeforeObjectAddedEvent
+from guillotina.interfaces import IBeforeObjectMovedEvent
 from guillotina.interfaces import IBeforeObjectRemovedEvent
 from guillotina.interfaces import IFileFinishUploaded
 from guillotina.interfaces import INewUserAdded
 from guillotina.interfaces import IObjectAddedEvent
+from guillotina.interfaces import IObjectDuplicatedEvent
 from guillotina.interfaces import IObjectLoadedEvent
+from guillotina.interfaces import IObjectLocationEvent
 from guillotina.interfaces import IObjectModifiedEvent
 from guillotina.interfaces import IObjectMovedEvent
 from guillotina.interfaces import IObjectPermissionsModifiedEvent
@@ -21,8 +24,8 @@ class ObjectEvent(object):
         self.object = object
 
 
-@implementer(IObjectMovedEvent)
-class ObjectMovedEvent(ObjectEvent):
+@implementer(IObjectLocationEvent)
+class ObjectLocationEvent(ObjectEvent):
     """An object has been moved"""
 
     def __init__(self, object, old_parent, old_name, new_parent, new_name, payload=None):
@@ -32,6 +35,16 @@ class ObjectMovedEvent(ObjectEvent):
         self.new_parent = new_parent
         self.new_name = new_name
         self.payload = payload
+
+
+@implementer(IObjectMovedEvent)
+class ObjectMovedEvent(ObjectLocationEvent):
+    """An object has been moved"""
+
+
+@implementer(IBeforeObjectMovedEvent)
+class BeforeObjectMovedEvent(ObjectLocationEvent):
+    pass
 
 
 class BaseAddedEvent(ObjectMovedEvent):
@@ -49,6 +62,13 @@ class BaseAddedEvent(ObjectMovedEvent):
 @implementer(IObjectAddedEvent)
 class ObjectAddedEvent(BaseAddedEvent):
     """An object has been added to a container"""
+
+
+@implementer(IObjectDuplicatedEvent)
+class ObjectDuplicatedEvent(ObjectAddedEvent):
+    def __init__(self, object, original_object, new_parent=None, new_name=None,
+                 payload=None):
+        super().__init__(object, new_parent, new_name, payload)
 
 
 @implementer(IBeforeObjectAddedEvent)